+++ src/main/java/be/baur/sdt/xpath/SDTFunctionContext.java
@@ -1,6 +1,11 @@
 package be.baur.sdt.xpath;
 
+import java.time.ZonedDateTime;
+
+import org.jaxen.Function;
 import org.jaxen.FunctionContext;
+import org.jaxen.SimpleFunctionContext;
+import org.jaxen.UnresolvableException;
 import org.jaxen.XPathFunctionContext;
 
 import be.baur.sdt.SDT;
@@ -25,55 +30,70 @@ import be.baur.sdt.xpath.function.TimestampFunction;
 import be.baur.sdt.xpath.function.TokenizeFunction;
 
 /**
- * A <code>FunctionContext</code> implementing the core XPath function library,
- * plus Jaxen extensions, and all SDT extensions.
+ * A <code>FunctionContext</code> providing the core XPath functions plus
+ * Jaxen extensions, all SDT extensions and evaluation metadata to support
+ * deterministic and/or context-dependent functions.
  * 
  * @see FunctionContext
  */
-public class SDTFunctionContext extends XPathFunctionContext
-{
-    private static final SDTFunctionContext instance = new SDTFunctionContext();
-
-	/**
-	 * Returns the default SDT function context. Do not extend this with your own
-	 * functions, as it will affect all XPath instances that use this global
-	 * default. Instead, extend <code>SDTFunctionContext</code> or create a new
-	 * instance to include your own functions.
-	 *
-	 * @return a function context
-	 */
-	public static SDTFunctionContext getInstance() {
-		return instance;
+public class SDTFunctionContext implements FunctionContext {
+	
+	// this class is essentially a wrapper around the default XPath context
+	private static final SimpleFunctionContext SFC = new XPathFunctionContext();
+	
+	// add SDT functions to the core Xpath functions and Jaxen extensions
+	static {
+		SFC.registerFunction(SDT.FUNCTIONS_NS_URI, CompareNumberFunction.NAME, new CompareNumberFunction());
+		SFC.registerFunction(SDT.FUNCTIONS_NS_URI, CompareStringFunction.NAME, new CompareStringFunction());
+		SFC.registerFunction(SDT.FUNCTIONS_NS_URI, CurrentDateTimeFunction.NAME, new CurrentDateTimeFunction());
+		SFC.registerFunction(SDT.FUNCTIONS_NS_URI, DateTimeFunction.NAME, new DateTimeFunction());
+		SFC.registerFunction(SDT.FUNCTIONS_NS_URI, DateTimeToLocalFunction.NAME, new DateTimeToLocalFunction());
+		SFC.registerFunction(SDT.FUNCTIONS_NS_URI, DateTimeToMillisFunction.NAME, new DateTimeToMillisFunction());
+		SFC.registerFunction(SDT.FUNCTIONS_NS_URI, DateTimeToTimeZoneFunction.NAME, new DateTimeToTimeZoneFunction());
+		SFC.registerFunction(SDT.FUNCTIONS_NS_URI, DocumentNodeFunction.NAME, new DocumentNodeFunction());
+		SFC.registerFunction(SDT.FUNCTIONS_NS_URI, FormatDateTimeFunction.NAME, new FormatDateTimeFunction());
+		SFC.registerFunction(SDT.FUNCTIONS_NS_URI, LeftFunction.NAME, new LeftFunction());
+		SFC.registerFunction(SDT.FUNCTIONS_NS_URI, MillisToDateTimeFunction.NAME, new MillisToDateTimeFunction());
+		SFC.registerFunction(SDT.FUNCTIONS_NS_URI, ParseDateTimeFunction.NAME, new ParseDateTimeFunction());
+		SFC.registerFunction(SDT.FUNCTIONS_NS_URI, ParseSDAFunction.NAME, new ParseSDAFunction());
+		SFC.registerFunction(SDT.FUNCTIONS_NS_URI, RenderSDAFunction.NAME, new RenderSDAFunction());
+		SFC.registerFunction(SDT.FUNCTIONS_NS_URI, RightFunction.NAME, new RightFunction());
+		SFC.registerFunction(SDT.W3CFUNCTIONS_NS_URI, StringJoinFunction.NAME, new StringJoinFunction());
+		SFC.registerFunction(SDT.FUNCTIONS_NS_URI, SystemTimeZoneFunction.NAME, new SystemTimeZoneFunction());
+		SFC.registerFunction(SDT.FUNCTIONS_NS_URI, TimestampFunction.NAME, new TimestampFunction());
+		SFC.registerFunction(SDT.FUNCTIONS_NS_URI, TokenizeFunction.NAME, new TokenizeFunction());
 	}
 
- 
+
+	private ZonedDateTime currentDateTime = null; // current context date and time
+
+
 	/**
-	 * Create a new XPath function context including the core XPath function
-	 * library, plus Jaxen extensions, and all SDT extensions.
+	 * Create a new SDT function context.
 	 */
-	public SDTFunctionContext() {
-		
-		super(true); // adds the core Xpath functions and Jaxen extensions
-		
-		registerFunction(SDT.FUNCTIONS_NS_URI, CompareNumberFunction.NAME, new CompareNumberFunction());
-		registerFunction(SDT.FUNCTIONS_NS_URI, CompareStringFunction.NAME, new CompareStringFunction());
-		registerFunction(SDT.FUNCTIONS_NS_URI, CurrentDateTimeFunction.NAME, new CurrentDateTimeFunction());
-		registerFunction(SDT.FUNCTIONS_NS_URI, DateTimeFunction.NAME, new DateTimeFunction());
-		registerFunction(SDT.FUNCTIONS_NS_URI, DateTimeToLocalFunction.NAME, new DateTimeToLocalFunction());
-		registerFunction(SDT.FUNCTIONS_NS_URI, DateTimeToMillisFunction.NAME, new DateTimeToMillisFunction());
-		registerFunction(SDT.FUNCTIONS_NS_URI, DateTimeToTimeZoneFunction.NAME, new DateTimeToTimeZoneFunction());
-		registerFunction(SDT.FUNCTIONS_NS_URI, DocumentNodeFunction.NAME, new DocumentNodeFunction());
-		registerFunction(SDT.FUNCTIONS_NS_URI, FormatDateTimeFunction.NAME, new FormatDateTimeFunction());
-		registerFunction(SDT.FUNCTIONS_NS_URI, LeftFunction.NAME, new LeftFunction());
-		registerFunction(SDT.FUNCTIONS_NS_URI, MillisToDateTimeFunction.NAME, new MillisToDateTimeFunction());
-		registerFunction(SDT.FUNCTIONS_NS_URI, ParseDateTimeFunction.NAME, new ParseDateTimeFunction());
-		registerFunction(SDT.FUNCTIONS_NS_URI, ParseSDAFunction.NAME, new ParseSDAFunction());
-		registerFunction(SDT.FUNCTIONS_NS_URI, RenderSDAFunction.NAME, new RenderSDAFunction());
-		registerFunction(SDT.FUNCTIONS_NS_URI, RightFunction.NAME, new RightFunction());
-		registerFunction(SDT.W3CFUNCTIONS_NS_URI, StringJoinFunction.NAME, new StringJoinFunction());
-		registerFunction(SDT.FUNCTIONS_NS_URI, SystemTimeZoneFunction.NAME, new SystemTimeZoneFunction());
-		registerFunction(SDT.FUNCTIONS_NS_URI, TimestampFunction.NAME, new TimestampFunction());
-		registerFunction(SDT.FUNCTIONS_NS_URI, TokenizeFunction.NAME, new TokenizeFunction());
+	public SDTFunctionContext() {}
+
+
+	@Override
+	public Function getFunction(String namespaceURI, String prefix, String localName) throws UnresolvableException {
+		return SFC.getFunction(namespaceURI, prefix, localName);
 	}
 
+
+	// Supporting methods for deterministic and/or context-dependent functions
+
+
+	/**
+	 * Returns the current system's clock date and time (with time zone) for this
+	 * context. The result is deterministic; multiple invocations will return the
+	 * same result.
+	 * 
+	 * @return a zoned date time
+	 */
+	public ZonedDateTime getCurrentDateTime() {
+
+		if (currentDateTime == null)
+			currentDateTime = ZonedDateTime.now();
+		return currentDateTime;
+	}
 }